// apps/question-server-nestjs/src/question/schemas/question.schemas.ts
/* eslint-disable prettier/prettier */
import { Prop, Schema, SchemaFactory } from "@nestjs/mongoose";
import { HydratedDocument } from "mongoose";

export type QuestionDocument = HydratedDocument<Question>

@Schema({
    timestamps: true
})
export class Question {
    @Prop({ required: true })
    title: string;

    @Prop()
    desc: string;

    @Prop()
    css: string;

    @Prop({ default: false })
    isPublished: boolean

    @Prop({ default: false })
    isStar: boolean

    @Prop({ default: false })
    isDeleted: boolean

    @Prop({ required: true })
    author: string;

    @Prop()
    componentList: {
        fe_id: string;
        type: string;
        title: string;
        isHidden: boolean;
        isLocked: boolean;
        props: object
    }[];
}

export const QuestionSchema = SchemaFactory.createForClass(Question);